f-nav: redesign header and navbar
