--- src/app/components/Footer/Footer.scss
@@ -1,62 +1,0 @@
-// SCSS base
-@import "../../App.scss";
-
-.footer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  height: 5.5em;
-
-  .footerContact {
-    text-transform: uppercase;
-    font-size: 1.1em;
-    line-height: 1.1em;
-
-    transform: rotate(-90deg);
-    width: fit-content;
-    margin-left: -2.2em;
-
-    a {
-      color: $color-3d-black;
-      margin: 1em;
-      position: relative;
-
-      &:visited,
-      &:link,
-      &:active {
-        text-decoration: none;
-      }
-    }
-  }
-
-  .footerBox {
-    height: 80%;
-
-    .footerLanguage {
-      font-family: -apple-system, BlinkMacSystemFont, "Segoe UI";
-      font-size: 0.8em;
-
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-
-      p {
-        &:hover {
-          color: gray;
-          text-decoration: none;
-          cursor: pointer;
-        }
-      }
-
-      span {
-        margin: 0.5em;
-      }
-    }
-    .footerSocial {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-  }
-}
