+++ src/app/components/Header/Header.scss
@@ -5,32 +5,20 @@
   display: flex;
   flex-direction: row;
   justify-content: space-between;
-  align-items: center;
+  align-items: flex-start;
 
   .headerLogo {
     display: flex;
     align-items: center;
     justify-content: center;
     img {
-      width: 3em;
+      width: 2.8em;
     }
   }
 
-  .headerNav {
+  .headerSocial {
     display: flex;
-    flex-direction: column;
-    align-items: flex-end;
     justify-content: center;
-
-    text-transform: uppercase;
-    font-size: 1.1em;
-    line-height: 1.1em;
-
-    .headerNavText {
-      position: relative;
-      color: $color-3d-black;
-
-      margin: 0.3em 0;
-    }
+    align-items: center;
   }
 }
