+++ src/app/components/Navbar/Navbar.scss
@@ -1,0 +1,2 @@
+// SCSS base
+@import "../../App.scss";
