+++ src/app/pages/Home/Home.tsx
@@ -2,14 +2,14 @@ import * as React from "react";
 import "./Home.scss";
 
 import { Header } from "../../components/Header/Header";
-import { Footer } from "../../components/Footer/Footer";
+import { Navbar } from "../../components/Navbar/Navbar";
 
 export const Home = (): JSX.Element => {
   return (
     <>
       <div className="background"></div>
       <Header />
-      <Footer />
+      <Navbar />
     </>
   );
 };
